app2/src/requirements.txt
alpaca-trade-api>=3.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
pytz>=2021.1
matplotlib>=3.4.0
backtrader>=1.9.76.123
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
aiofiles>=0.8.0
websockets>=10.0
python-multipart>=0.0.5
pytest>=7.0.0
requests>=2.0.0
httpx>=0.20.0
pytest-mock>=3.10.0